Add option to deploy without settings_data.json
Sometimes I just want to upload all new files without overwriting
my current settings. It's easy to hop into the theme editor in
the Shopify backend and rollback to a previous version of the
file but this makes it even easier.

Simply pass `--no-json` as an option to the `shopify:upload` task:

    grunt shopify:upload --no-json